requirements.txt
# Core Deep Learning
torch>=2.5.1
torchvision>=0.20.1
ultralytics>=8.3.221

# Computer Vision
opencv-python>=4.10.0
numpy>=1.26.0
Pillow>=10.0.0

# GUI Framework
PySide6>=6.8.0
PySide6-Addons>=6.8.0
PySide6-Essentials>=6.8.0

# Configuration
PyYAML>=6.0

# Utilities
tqdm>=4.65.0
psutil>=5.9.0
